Improve the full-repository config docs, and add reference names.
The general documentation on configuring repositories previously was
somewhat dense. A couple of paragraphs that highlighted a bunch of
important, but somewhat buried, information, and a paragraph that had no
business being there (likely a copy/paste), since it really belonged to
the auto-fetching section.

Neither section broke down the configuration options, or really guided
users toward setting up API tokens. They expected a bit of knowledge
that users may lack.

This change redoes the documentation, breaking it up into less-dense
paragraphs, specifically calling out the supported repository types, and
providing a list of each option for the worker configuration and what's
allowed or expected for both ways of configuring repositories.

An "important" callout box was added for the information on the
repository configuration on the Review Board side, for auto-fetching of
repositories, to help make this really stand out. Example scenarios and
a pointer to RB Gateway was added. The copy of the original version of
this information was removed from the manual section.

Section headers have also been renamed to help differentiate the two,
and reference names have been added to some sections in the file.

Testing Done:
Built the docs and read through them. Checked for spelling errors, bad
links, and build errors.

Reviewed at https://reviews.reviewboard.org/r/11527/