Add release notes for Review Bot 2.0.1.
This is a small release with bug fixes for Python 3 and Review Board
4.0.

The plan initially was to release these fixes in Review Bot 3.0, but
that release is undergoing a lot of additional work, and we don't want
to delay these fixes any longer.

Testing Done:
Built the docs. Verified there were no build errors or spelling errors.

Reviewed at https://reviews.reviewboard.org/r/11561/