Update cpplint for new tool support and improve output parsing.
This updates `CPPLintTool` to subclass the newer `BaseTool`, and to use
the newer features it offers (dependencies, improved file extension
handling, centralized base command building, and standardized error code
output).

The new version of the tool has less verbose, safer output parsing. Both
the old and new versions used a regex, but the new one is designed to
better fit with the output from the tool, and to avoid a bunch of
now-unnecesary processing.

Testing Done:
Unit tests pass for Python 2.7 and 3.x.

Reviewed at https://reviews.reviewboard.org/r/11567/