Remove support for Buildbot.
We've had some preliminary support for Buildbot in Review Bot for the
past couple major releases, but it was never more than a prototype. It
leveraged `buildbot try` in order to run a build, but we had several
issues using this for anything real:

1. It made assumption about patch levels for diffs (which must be passed
   to `buildbot `try`), and those assumptions made this useless for
   non-Git-style diffs.
2. It didn't calculate or request any options that would be required in
   order to run changes from other SCMs, like Subversion or Perforce.
3. It may not have even really worked with Git, outside of limited
   setups (due to issues locating the repository).
4. There's no useful output that comes from `buildbot try`, aside from
   some debug information and a general "pass" or "fail" response, all
   of which ends up in the review body. We had no way to even get a link
   to the build.

The plan is to add formal Buildbot support as an integration in the
future, alongside CircleCI, Jenkins, and Travis-CI. For now, though,
we're scrapping this so that people don't end up getting frustrated
trying to configure a tool that won't work for them.

The old documentation page will remain, but won't be included in any
Table of Contents. This will help explain where the support went, and
direct people to an alternative solution.

Testing Done:
Unit tests pass.

Built the docs. Verified the page wasn't in the table of contents, but
was available, and wasn't generating any warnings. Tested the links within.

Reviewed at https://reviews.reviewboard.org/r/11586/