Allow tool unit tests to provide multiple files to test against.
This adds an `other_files` argument to `run_tool_execute()` that
specifies a mapping of filenames to file contents. These files will be
written to the tree, allowing the tool to access those files. This is
mostly useful for full-repository unit tests.

When provided, the returned tuple will contain a mapping of filenames to
`File` entries, rather than a single `File` entry.

Testing Done:
Used this along with the upcoming unit tests for Cargo.

Reviewed at https://reviews.reviewboard.org/r/11653/