Rename Credential Checker to Secret Scanner.
The Credential Checker used to be a fairly simple scanner for common
types of credentials, but has since expanded in scope a bit. To give the
tool a broader purpose, and to better align with terminology from other
tools on the market, we're renaming this to Secret Scanner.

Testing Done:
Unit tests pass.

Re-installed the package and verified that `reviewbot worker` picked up
the new tool.

Reviewed at https://reviews.reviewboard.org/r/11655/