Update the documentation for Go Tool.
This fleshes out the docs, listing the file types that are supported and
adding installation instructions for the tool. It details the
configuration options and points to the necessary docs on
full-repository support.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and bad links.

Reviewed at https://reviews.reviewboard.org/r/11783/