Add a list of supported file types to the PMD docs.
This expands upon the PMD docs, providing a list of file types that are
supported, as well as relevant file extensions. A handy list of URLs to
Java implementations for each file type in PMD (each containing the list
of file extensions) is provided to make future maintenance on this file
type list a bit easier.

Some wording in the Configuration section was also changed for
consistency.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and bad links.

Reviewed at https://reviews.reviewboard.org/r/11784/