Update the documentation for the new Secret Scanner.
This updates the docs for Secret Scanner to rename it from Credential
Checker, to match the format used for all the other new tool docs, and
to list all the types of secrets that are checked.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and bad links.

Reviewed at https://reviews.reviewboard.org/r/11785/