Skip patching symlinks, binary files, or unknown file changes.
Review Bot was attempting to apply patches to symlinks and to binary
files, but had no means of doing so. For the kinds of tools that are run
by Review Bot, we shouldn't usually care much about these types of
changes.

An argument can be made for symlinks, but we're currently unable to
guarantee we'll apply these right, and symlink target support hasn't
been available until Review Board 4.0.6 anyway.

This change skips these file types, and skips unknown FileDiff status
codes, in order to minimize the kinds of changes that need to be patched
to the tree or passed to tools.

Testing Done:
Relevant unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12072/