Add the beginnings of unit test support for the Review Bot extension.
This adds a basic test runner and `TestCase` class for building unit
tests for the Review Bot extension.

Currently, this uses nose, which is not really the direction we want to
go. However, we previously had some issues with getting pytest working
on Review Board 4/Python 2, so for now we're using nose, with the plan
to move to pytest in the future.

Testing Done:
Tested with upcoming changes adding unit tests.

Reviewed at https://reviews.reviewboard.org/r/12092/