Fix refreshing the list of tools.
When refreshing the list of tools, we attempted to render a template to
send back to the API call. This was passing a legacy parameter to
`RequestContext`, which no longer works, causing the refresh spinner to
spin forever and for a traceback to appear in reviewboard.log.

This fortunately did not impede the refresh operation, but it did give
the appearance that refresh had failed.

This change fixes that up, removing the template and just returning a
hard-coded result.

Testing Done:
Tested refreshing tools. It successfully refreshed and reloaded the page.

Reviewed at https://reviews.reviewboard.org/r/12062/