Don't error out if a FBInfer run doesn't generate a report.
The FBInfer post-processing code made the assumption that reports would
always be provided after a run. This doesn't appear to always be the
case in production.

The tool now checks for the presence of a report and returns empty
results if not found.

Testing Done:
Unit tests pass.

Tested with a real FBInfer run that generated no results. Verified
no errors were logged.

Reviewed at https://reviews.reviewboard.org/r/12138/