Modernize the documentation and add instructions on Docker.
This makes some final changes to the documentation before the release of
Review Bot 3. The main page of the docs now contain:

1) General information on Review Bot
2) Guided information on installation, configuration, and upgrading.
3) All tools, grouped by language or category.

The installation docs have been updated to cover both manual
installation (same guide we had before) and the new Docker images. The
Docker instructions are based on the Review Board Docker instructions,
and cover the range of pre-built images, general setup and usage, and
instructions on how to create a custom `Dockerfile` with the explicit
list of tools needed.

The docs on upgrading have been cleaned up a little bit to add separate
sections on upgrading the extension and the workers.

Many anchors have been added, and some have been moved to point to the
new instructions on the main page.

The `README.rst` has also been updated to reflect some of the new
content shown on the main docs page.

Testing Done:
Built the docs, checking for link issues or build errors.

Read through the content, trying to find any issues or spelling errors.

Reviewed at https://reviews.reviewboard.org/r/12139/