Display Python and Celery info and deprecation warnings on startup.
When starting the worker, we now check the Python version. If it's
Python 2, we warn that Python 2 and Celery 3 are no longer supported and
may have security risks, recommending an upgrade to Python 3.

If running on Python 3, we show the Python and Celery versions (as a
reference) and have a general recommendation to keep up with versions.

Testing Done:
Tested running under Python 2 and 3, checking the resulting output
and version numbers.

Reviewed at https://reviews.reviewboard.org/r/12158/