Switch all console sessions in the docs to the console lexer.
This updates the documentation to use the `console` lexer for all sample
terminal sessions. This will display with our terminal component on the
website, and will highlight correctly.

There's also a fix for an error around a code block that was causing the
content to appear within a quote block.

Testing Done:
Checked every terminal in every page in the docs to ensure the appearance
was correct.

Reviewed at https://reviews.reviewboard.org/r/12216/