Fix the links to the Celery documentation.
The URL for the Celery site has changed. For a while, the old domain was
unavailable, though they've since reclaimed it and redirected it to the
new domain.

To be on the safe side, we now use the new domain in all of our links.

Testing Done:
Checked all the links to ensure they worked.

Reviewed at https://reviews.reviewboard.org/r/12382/