Fix a regression with running jshint.
Review Bot 3 introduced a new jshint implementation that made use of a
helper file to output as JSON.

Unfortunately, the helper file did not ship with the package, due to the
lack of `include_package_data=True` in the package configuration and
the lack of a `MANIFEST.in` file.

This fixes the packaging, ensuring we now ship the missing file.

Testing Done:
Built packages and verified that the helper script, and other files from
our source tree, were in the resulting wheel and source distributions.

Reviewed at https://reviews.reviewboard.org/r/12433/