Cap the version of Celery to avoid regressions with Python 3.6.
Celery 5.3 alpha is now up on PyPI, and this is not compatible with
Python 3.6. Technically, Celery 5.2.x wasn't compatible either, but it
did seem to run.

As Celery is apparently not using semantic versioning, we can't trust
that 5.x will remain compatible. We're now capping the version to the
5.1.x range for Python 3.6, and 5.2.x for newer versions.

It also bumps up the minimum version for each range.

Testing Done:
Successfully installed packages on each version of Python and ran
unit tests for the extension and worker.

Reviewed at https://reviews.reviewboard.org/r/12455/