Cap the version of the kombu library on Python 3.6.
Celery 5.1.x (required for Python 3.6) depends on `kombu>=5.1.0,<6.0`.
The problem is that they've dropped Python 3.6 support in kombu 5.2,
meaning that anyone trying to install this range of Celery versions will
end up with a broken environment.

We try to help them out (and help our CI out) by capping the version
range of kombu so that only versions compatible with Python 3.6 will
install.

Testing Done:
Installed packages on Python 3.6 and verified I got the right version
of kombu (and that it removed the incompatible version).

Unclear at this stage if there are other bad dependency ranges that
will have to be capped.

Reviewed at https://reviews.reviewboard.org/r/12459/