Fix issues running unit tests with newer RBTools and with writable root.
Recent RBTools 3.x releases updated API error messages to be a bit more
useful. This broke a couple unit tests that checked for the resulting
error message.

These tests have been updated to check the current strings, and
`dev-requirements.txt` has been updated to specifically require the
newer versions.

We had a couple other unit tests that checked for issues writing a
cookie file. They attempted to point to a directory off of `/` that it
couldn't write to. This assumed the tests were being run in an
environment without a writable `/`, but that's not the case in CI
environments.

These tests now base their paths off of `/dev/null`, which should convey
the same sort of error.

Testing Done:
Unit tests pass with the newer RBTools.

Attempted to simulate the writable root test in a CI environment. The
real test will happen after this has landed.

Reviewed at https://reviews.reviewboard.org/r/12460/