Cap the maximum number of lines for a comment to 10.
Some tools include a pretty wide range of lines when reporting a
comment. PMD, for instance, will list the entire line range for a
function or a comment, which is excessive and not particularly useful.

We now cap comments to 10 lines. When capped, the total line range is
shown in the comment, giving people enough context to see the entire set
of lines for the reported issue.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12986/