Cap the kombu and RBTools dependencies on Python < 3.8.
Kombu is dropping Python 3.7 support in their upcoming 5.3 releases,
which are available on PyPI right now as pre-releases. When installing
Review Bot via eggs, `easy_install` doesn't know to avoid pre-releases
(or to consider supported Python version metadata), so we have to cap
this until we drop support for Python 2.7 and can depend fully on
wheels.

RBTools dropped Python 2.7 and 3.6 in RBTools 4. We had a pretty old
minimum version, so I've taken the opportunity to bump that up. We now
depend on RBTools 3.x for Python 2.7/3.7, and RBTools 4.x or higher for
newer releases.

Testing Done:
Successfully installed into new Python 2.7, 3.7, and 3.8 environments.

Reviewed at https://reviews.reviewboard.org/r/13081/