Cap the version of importlib-metadata on Python <= 3.7.
Celery 5.2.7 fails to import on Python 3.7 environments when
`importlib-metadata` >= 5.0 is installed. The `celery.Celery` import
fails.

To resolve this, we need to cap `importlib-metadata` on older versions
of Python.

Testing Done:
Successfully installed on a Python 3.7 environment and verified that
`celery.Celery` could be imported.

Reviewed at https://reviews.reviewboard.org/r/13083/