Add support for PMD 7.
PMD just released a big new major version, 7.0.0. This makes a fairly
large change to the basic command-line UI, but is mostly otherwise
compatible.

Some of the output is slightly different for a few of the tests. I've
changed the simulation tests to simulate PMD7, and the integration tests
sometimes need to conditionalize the expected output. We also had one
test (for a config file with no rules) that was obsolete even in PMD
6.55, which treated a file with no active rules as a warning rather than
an error.

Testing Done:
Ran unit tests with both PMD 6.55.0 and 7.0.0. Saw that everything
passed.

Reviewed at https://reviews.reviewboard.org/r/13710/