Update pyflakes output parsing and CargoTool tests for latest versions.
Since the last Review Bot release, PyFlakes has updated to have new
error handling for unparsable files when run on newer versions of Python
(3.11.4+). The new error information includes a line number, but no
column number. We now handle parsing this information correctly, and
updated unit tests to check for it. As this is really only used for
things like null byte encoding issues in a file, it's not a major issue
in practice.

CargoTool also had numerous changes to the output format, but this
didn't impact the actual tool, just the unit tests. We now check for
older and newer variations of the output in the tests.

Testing Done:
Unit tests passed on my local dev server with the older dependencies
and in CI with newer dependencies.

Reviewed at https://reviews.reviewboard.org/r/14100/